29/07 02:30h

+++ python3_Exercicios/ex010.py
@@ -1,2 +1,2 @@
-rs = int(input('R$= '))
+rs = float(input('R$= '))
 print('With R$:{:.2f} = U$:{:.2f} dollars'.format(rs, (rs/5.34)))
